Fixed files on it's correct places and added main/status and correct files to positions
