--- experimental/notebooks/Yolo-V8/PL01.csv
@@ -1,14 +1,0 @@
-Parking Slot,Point1_X,Point1_Y,Point2_X,Point2_Y,Point3_X,Point3_Y,Point4_X,Point4_Y
-Box 1,1092,675,1141,565,1037,531,967,624
-Box 2,963,618,1032,527,946,485,869,567
-Box 3,863,565,935,488,854,454,783,535
-Box 4,777,532,847,454,793,446,708,505
-Box 5,704,500,784,446,737,433,649,478
-Box 6,643,476,724,434,686,418,595,457
-Box 7,591,453,679,417,650,401,550,436
-Box 8,546,433,641,399,595,391,512,419
-Box 9,509,416,588,389,532,384,469,407
-Box 10,443,409,477,380,416,381,380,414
-Box 11,375,411,410,379,371,376,326,411
-Box 12,321,409,364,376,317,382,259,420
-Box 13,248,424,311,380,242,393,181,435
