+++ src/data/PL02.csv
@@ -1,0 +1,24 @@
+Parking Slot,Point1_X,Point1_Y,Point2_X,Point2_Y,Point3_X,Point3_Y,Point4_X,Point4_Y
+Box 1,1040,703,1058,601,1166,670,1167,718
+Box 2,1033,705,1053,599,965,543,932,625
+Box 3,927,624,962,540,888,493,844,564
+Box 4,840,559,885,491,821,454,777,510
+Box 5,774,506,819,450,769,416,718,466
+Box 6,713,463,765,413,722,386,667,429
+Box 7,665,424,718,384,680,361,627,397
+Box 8,623,392,677,357,641,338,588,368
+Box 9,586,366,637,335,610,318,556,344
+Box 10,554,341,607,315,580,297,529,323
+Box 11,526,320,577,295,554,281,503,303
+Box 12,502,300,552,278,534,265,482,287
+Box 13,479,284,532,263,511,250,462,271
+Box 14,459,269,508,247,489,239,439,258
+Box 15,439,255,486,236,468,228,427,243
+Box 16,459,719,555,626,509,568,387,675
+Box 17,387,667,507,563,470,512,361,596
+Box 18,358,591,466,509,436,470,331,539
+Box 19,329,533,429,467,410,430,308,490
+Box 20,308,485,405,428,382,401,297,447
+Box 21,378,397,293,441,286,405,364,371
+Box 22,288,400,361,365,342,346,276,378
+Box 23,269,371,337,343,315,330,257,354
