+++ src/data/PL03.csv
@@ -1,0 +1,6 @@
+Parking Slot,Point1_X,Point1_Y,Point2_X,Point2_Y,Point3_X,Point3_Y,Point4_X,Point4_Y
+Box 1,290,514,347,507,401,556,349,571
+Box 2,359,576,412,559,542,634,458,658
+Box 3,466,663,543,641,676,696,604,719
+Box 4,580,476,628,470,712,495,674,507
+Box 5,679,511,717,498,786,518,763,539
