+++ src/data/parking_status.csv
@@ -1,0 +1,2 @@
+ParkingLotID,Timestamp,SP1,SP2,SP3,SP4,SP5,SP6,SP7,SP8,SP9,SP10,SP11,SP12,SP13
+PL01,2024-07-24 17:06:07,empty,occupied,occupied,occupied,empty,empty,empty,empty,occupied,empty,empty,empty,empty
